Crate ttl_cache

This crate provides a time sensitive key-value cache. When an item is inserted it is given a TTL. Any value that are in the cache after their duration are considered invalid and will not be returned on lookups.

Structs

Iter

IterMut

OccupiedEntry
A view into a single occupied location in the cache that was unexpired at the moment of lookup.

TtlCache
A time sensitive cache.

VacantEntry
A view into a single empty location in the cache

Enums

Entry
A view into a single location in a map, which may be vacant or occupied.